requirements-windows-gpu.txt
# Windows GPU-friendly CosyVoice dependencies for this machine.
# We intentionally do not pin torch/torchaudio here, because this PC
# already has a working Python 3.11 + torch 2.5.1+cu121 stack.

conformer==0.3.2
diffusers==0.29.0
fastapi==0.115.6
fastapi-cli==0.0.4
gdown==5.1.0
gradio==5.4.0
grpcio==1.57.0
grpcio-tools==1.57.0
hydra-core==1.3.2
HyperPyYAML==1.2.3
inflect==7.3.1
librosa==0.10.2
lightning==2.2.4
matplotlib==3.7.5
modelscope==1.20.0
networkx==3.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
# onnxruntime-gpu is installed separately from the official CUDA 12 feed
# in scripts/setup_gpu_env.ps1 so it matches the local torch CUDA stack.
openai-whisper==20231117
protobuf==4.25
pyarrow==18.1.0
pydantic==2.7.0
pyworld==0.3.4
rich==13.7.1
soundfile==0.12.1
tensorboard==2.14.0
transformers==4.51.3
x-transformers==2.11.24
uvicorn==0.30.0
wetext==0.0.4
wget==3.2
